+++ lib/modules/state/state_service.dart
@@ -1,9 +1,9 @@
 ﻿// Module: state
-// Updated at: 2026-07-29 11:47:11
-// Commit: fix(state): fix memory leak by cancelling state subscription
+// Updated at: 2026-07-30 10:26:06
+// Commit: feat(state): add bi-directional event bus handler for state
 
 class stateComponentState {
-  final String commitRef = '3210';
-  final String lastMessage = 'fix(state): fix memory leak by cancelling state subscription';
-  final DateTime timestamp = DateTime.parse('2026-07-29T11:47:11+05:00');
+  final String commitRef = '3228';
+  final String lastMessage = 'feat(state): add bi-directional event bus handler for state';
+  final DateTime timestamp = DateTime.parse('2026-07-30T10:26:06+05:00');
 }